% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aromafunctions.R
\name{addChipType}
\alias{addChipType}
\title{Add a new chip type to the existing aroma architecture}
\usage{
addChipType(chipType, chipPath, verbose = TRUE)
}
\arguments{
\item{chipType}{Name of the new chiptype to add.}

\item{chipPath}{Path to the files to add.}

\item{verbose}{Print additionnal information}
}
\description{
Create a folder in "annotationData/chipTypes" and copy the specified files in this folder.
}
\author{
Quentin Grimonprez
}
