% Generated by roxygen2 (4.0.0): do not edit by hand
\name{CNAobjectToCGHcallObject}
\alias{CNAobjectToCGHcallObject}
\title{Convert CNAobject}
\usage{
CNAobjectToCGHcallObject(CNAobject)
}
\arguments{
\item{CNAobject}{Output object of segment function from DNAcopy package}
}
\value{
a list at the required format of \link{callingProcess}.
}
\description{
convert CNA object (output of the function segment from DNAcopy package) into a list for the argument segmentData of the function \link{callingProcess}.
}
\author{
Quentin Grimonprez
}
\seealso{
callingProcess
}

