% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cghcall.R
\name{segmentationObject}
\alias{segmentationObject}
\title{Create the list of parameters for \link{segmentation} function}
\usage{
segmentationObject(copynumber, chromosome, position, featureNames, sampleNames)
}
\arguments{
\item{copynumber}{A vector containing the copy-number signal for one patient and one chromosome.}

\item{chromosome}{Chromosome associated with the copy-number signal.}

\item{position}{Position of the signal.}

\item{featureNames}{Names of the probes (not necessary).}

\item{sampleNames}{Name of the sample (not necessary).}
}
\value{
a list in the right format for \link{segmentation} function
}
\description{
create the list of parameters for \link{segmentation} function
}
\author{
Quentin Grimonprez
}
