\name{addData}
\alias{addData}
\title{add a new data-set to the existing aroma architecture}
\usage{
addData(dataSetName, dataPath, chipType)
}
\arguments{
  \item{dataSetName}{Name of the new data-set to add.}

  \item{dataPath}{Path to the files to add.}

  \item{chipType}{Name of the chip used for the data.}
}
\description{
Create a folder in "rawData" and copy the specified files
in this folder.
}
\author{
Quentin Grimonprez
}

